requirements.txt
autopep8==1.4.4
backports.csv==1.0.7
bcrypt==3.1.7
beautifulsoup4==4.7.1
cairocffi==1.0.2
CairoSVG==2.4.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-allauth==0.39.1
django-countries==5.3.3
django-debug-toolbar==2.0
django-filebrowser-no-grappelli==3.7.8
django-htmlmin==0.11.0
django-import-export==1.2.0
django-recaptcha3==0.1.0
django-rest-framework==0.1.0
django-tinymce4-lite==1.7.5
djangorestframework==3.10.1
et-xmlfile==1.0.1
html5lib==1.0.1
idna==2.8
jdcal==1.4.1
jsmin==2.2.2
numpy==1.16.4
oauthlib==3.0.2
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.1.0
pycodestyle==2.5.0
pycparser==2.19
Pyphen==0.9.5
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
stripe==2.32.1
tablib==0.13.0
tinycss2==1.0.2
urllib3==1.25.3
WeasyPrint==48
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
